requirements.txt
fastapi==0.79.0
ffmpeg_python==0.2.0
jose==1.0.0
lz4==4.0.1
passlib==1.7.4
Pillow==9.2.0
pydantic==1.9.1
python_jose==3.3.0
SQLAlchemy==1.4.35
tqdm==4.64.0
typing_extensions==4.3.0
